# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginRequest:
    properties:
      identify:
        example: admin
        type: string
      password:
        example: admin
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: admin@gmail.com
        type: string
      name:
        example: Admin
        type: string
      password:
        example: admin
        type: string
      phone:
        example: "0984807356"
        type: string
      username:
        example: admin
        type: string
    required:
    - name
    - password
    - phone
    type: object
  category.CreateCategoryRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
  category.DeleteCategoriesRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  category.UpdateCategoryRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using email or username and password.
      parameters:
      - description: Login request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email, and password
      parameters:
      - description: Register data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success": "Register success"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: '{"error": "invalid request body"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: '{"error": "server error"}'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User register
      tags:
      - auth
  /categories:
    delete:
      consumes:
      - application/json
      description: Deletes multiple categories by their IDs
      parameters:
      - description: List of IDs
        in: body
        name: ids
        required: true
        schema:
          $ref: '#/definitions/category.DeleteCategoriesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete multiple categories
      tags:
      - categories
    get:
      description: Returns a list of categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get category list
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Creates a new category and returns the created category
      parameters:
      - description: Create data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/category.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    get:
      description: Returns a category by ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a category
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Updates a category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/category.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - categories
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"