# internal/modules/category/dto.go
package category

type CreateCategoryRequest struct {
	Name string `json:"name" validate:"required"`
	Slug string `json:"slug" validate:"required"`
}

type UpdateCategoryRequest struct {
	Name string `json:"name" validate:"required"`
	Slug string `json:"slug" validate:"required"`
}

type DeleteCategoriesRequest struct {
	IDs []int64 `json:"ids"`
}